small fixes

+++ src/store/modules/Authentication/mutations.ts
@@ -3,7 +3,6 @@ import { MutationTree } from "vuex"
 import { AuthenticationState } from "./index"
 import { AuthenticateResponseModel } from "@/models/Mediakiwi/Response/Authentication/AuthenticateResponseModel"
 import router from "@/router"
-import ls from "secure-ls"
 
 export type Mutations<S = AuthenticationState> = {
   [MutationTypes.SET_PROFILE](state: S, payload: AuthenticateResponseModel|null): void,
@@ -24,16 +23,22 @@ export const mutations: MutationTree<AuthenticationState> & Mutations = {
       }
     },
     [MutationTypes.AUTHENTICATE](state: AuthenticationState, payload: boolean): void {
-      state.isLoggedIn = true;
-      router.push("/");
+      if (payload) {
+        state.isLoggedIn = true;
+        router.push("/");
+      }
     },
     [MutationTypes.UNAUTHENTICATE](state: AuthenticationState, payload: boolean): void {
-      window.sessionStorage.clear();
-      window.localStorage.clear();
-      state.isLoggedIn = false;
-      router.push("/login");
+      if (payload) {
+        window.sessionStorage.clear();
+        window.localStorage.clear();
+        state.isLoggedIn = false;
+        router.push("/login");
+      }
     },
     [MutationTypes.RESET_PASSWORD](state:AuthenticationState, payload: boolean): void {
-      router.push("/");
+      if (payload) {
+        router.push("/");
+      }
     }
 }
